Split util helpers into focused modules (#1277)
