+++ custom_components/spook/ectoplasms/automation/repairs/unknown_area_references.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 from homeassistant.components import automation
 from homeassistant.helpers import area_registry as ar
 
+from ....entity_filtering import async_filter_known_area_ids, async_get_all_area_ids
 from ....repairs import AbstractSpookEntityComponentUnknownReferencesRepair
-from ....util import async_filter_known_area_ids, async_get_all_area_ids
 
 if TYPE_CHECKING:
     from typing import Any
