+++ custom_components/spook/ectoplasms/proximity/repairs/unknown_zone.py
@@ -8,8 +8,8 @@ from homeassistant.const import EVENT_COMPONENT_LOADED
 from homeassistant.helpers import entity_registry as er
 
 from ....const import LOGGER
+from ....entity_filtering import async_get_all_entity_ids
 from ....repairs import AbstractSpookRepair
-from ....util import async_get_all_entity_ids
 
 if TYPE_CHECKING:
     from homeassistant.components.proximity.coordinator import (
