+++ custom_components/spook/ectoplasms/script/repairs/unknown_floor_references.py
@@ -7,8 +7,8 @@ from homeassistant.helpers import floor_registry as fr
 from homeassistant.helpers.entity_component import DATA_INSTANCES, EntityComponent
 
 from ....const import LOGGER
+from ....entity_filtering import async_filter_known_floor_ids, async_get_all_floor_ids
 from ....repairs import AbstractSpookRepair
-from ....util import async_filter_known_floor_ids, async_get_all_floor_ids
 
 
 class SpookRepair(AbstractSpookRepair):
