+++ custom_components/spook/entity_filtering.py
@@ -1,10 +1,7 @@
-"""Spook - Your homie."""
+"""Spook - Your homie. Entity, registry, and template filtering helpers."""
 
 from __future__ import annotations
 
-import asyncio
-import importlib
-from pathlib import Path
 import re
 from typing import TYPE_CHECKING, Any
 
@@ -38,15 +35,12 @@ from homeassistant.helpers import (
 )
 from homeassistant.helpers.template import Template
 
-from .const import DOMAIN, LOGGER
+from .const import LOGGER
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterable, Sequence
-    from types import ModuleType
 
-    from homeassistant.config_entries import ConfigEntry
     from homeassistant.core import HomeAssistant
-    from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 
 # Entity domains to ignore when filtering unknown entities
@@ -221,98 +215,6 @@ def async_get_all_entity_ids(
     if include_all_none:
         return _CACHED_ALL_ENTITY_IDS.union({ENTITY_MATCH_ALL, ENTITY_MATCH_NONE})
     return _CACHED_ALL_ENTITY_IDS.copy()
-
-
-async def async_forward_setup_entry(
-    hass: HomeAssistant,
-    entry: ConfigEntry,
-) -> None:
-    """Set up Spook ectoplasms."""
-    LOGGER.debug("Setting up Spook ectoplasms")
-
-    modules: list[ModuleType] = []
-
-    def _load_all_ectoplasm_modules() -> None:
-        """Load all Spook ectoplasm modules."""
-        for module_file in Path(__file__).parent.rglob("ectoplasms/*/__init__.py"):
-            module_path = str(module_file.relative_to(Path(__file__).parent))[
-                :-3
-            ].replace(
-                "/",
-                ".",
-            )
-            LOGGER.debug("Loading Spook ectoplasm: %s", module_path)
-            module = importlib.import_module(f".{module_path}", __package__)
-            if hasattr(module, "async_setup_entry"):
-                modules.append(module)
-                LOGGER.debug("Setting up Spook ectoplasm: %s", module_path)
-
-    await hass.async_add_import_executor_job(_load_all_ectoplasm_modules)
-    await asyncio.gather(*(module.async_setup_entry(hass, entry) for module in modules))
-
-
-async def async_forward_platform_entry_setups_to_ectoplasm(
-    hass: HomeAssistant,
-    entry: ConfigEntry,
-    async_add_entities: AddEntitiesCallback,
-    platform: Platform,
-) -> None:
-    """Set up Spook ectoplasm platform."""
-    LOGGER.debug("Setting up Spook ectoplasm platform: %s", platform)
-
-    modules: list[ModuleType] = []
-
-    def _load_all_ectoplasm_platform_modules() -> None:
-        """Load all Spook ectoplasm platform modules."""
-        for module_file in Path(__file__).parent.rglob(f"ectoplasms/*/{platform}.py"):
-            module_path = str(module_file.relative_to(Path(__file__).parent))[
-                :-3
-            ].replace(
-                "/",
-                ".",
-            )
-            LOGGER.debug("Loading Spook %s from ectoplasm: %s", platform, module_path)
-            modules.append(importlib.import_module(f".{module_path}", __package__))
-            LOGGER.debug("Setting up Spook ectoplasm %s: %s", platform, module_path)
-
-    await hass.async_add_import_executor_job(_load_all_ectoplasm_platform_modules)
-    await asyncio.gather(
-        *(
-            module.async_setup_entry(hass, entry, async_add_entities)
-            for module in modules
-        )
-    )
-
-
-def link_sub_integrations(hass: HomeAssistant) -> bool:
-    """Link Spook sub integrations."""
-    LOGGER.debug("Linking up Spook sub integrations")
-
-    changes = False
-    for manifest in Path(__file__).parent.rglob("integrations/*/manifest.json"):
-        LOGGER.debug("Linking Spook sub integration: %s", manifest.parent.name)
-        dest = Path(hass.config.config_dir) / "custom_components" / manifest.parent.name
-        if not dest.exists():
-            src = (
-                Path(hass.config.config_dir)
-                / "custom_components"
-                / DOMAIN
-                / "integrations"
-                / manifest.parent.name
-            )
-            dest.symlink_to(src)
-            changes = True
-    return changes
-
-
-def unlink_sub_integrations(hass: HomeAssistant) -> None:
-    """Unlink Spook sub integrations."""
-    LOGGER.debug("Unlinking Spook sub integrations")
-    for manifest in Path(__file__).parent.rglob("integrations/*/manifest.json"):
-        LOGGER.debug("Unlinking Spook sub integration: %s", manifest.parent.name)
-        dest = Path(hass.config.config_dir) / "custom_components" / manifest.parent.name
-        if dest.exists():
-            dest.unlink()
 
 
 @callback
