+++ custom_components/spook/select.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from homeassistant.const import Platform
 
-from .util import async_forward_platform_entry_setups_to_ectoplasm
+from .setup_helpers import async_forward_platform_entry_setups_to_ectoplasm
 
 if TYPE_CHECKING:
     from homeassistant.config_entries import ConfigEntry
